handle scripts with attached signatures
POC only, code's pretty gross. When pipethis downloads a script, it checks for
a clearsigned block before anything else. If it finds one, it saves the signed
message and the signature separately. If it doesn't find one, it saves the
whole thing as usual.

This gets a little weird because when Go decodes a clearsigned block into the
signed message and the signature, it replaces all the LFs with CRLFs. That's
fine according to the spec (thanks @dgryski!), but the Bash shell doesn't
understand CRs. I'm working around that by stripping all the CRs off the signed
message. I'm not at all sure that's kosher, but it seems to work!

see #7